Organize sandbox datasets under data folder. Add small dataset
