+++ sandbox/data/small/small.config.js
@@ -1,0 +1,1 @@
+module.exports = {};
