+++ sandbox/data/small/small.data.js
@@ -1,0 +1,34 @@
+module.exports = {
+    links: [
+        {
+            source: 1,
+            target: 2
+        },
+        {
+            source: 1,
+            target: 3
+        },
+        {
+            source: 1,
+            target: 4
+        }
+    ],
+    nodes: [
+        {
+            id: 1,
+            name: 'Node 1'
+        },
+        {
+            id: 2,
+            name: 'Node 2'
+        },
+        {
+            id: 3,
+            name: 'Node 3'
+        },
+        {
+            id: 4,
+            name: 'Node 4'
+        }
+    ]
+};
